change number value from Double to Long, add db config

+++ src/main/java/com/example/test_task/model/NumberEntity.java
@@ -16,7 +16,7 @@ public class NumberEntity {
     private Long id;
 
     @Column("NUMBER_VALUE")
-    private Double numberValue;
+    private Long numberValue;
 
 }
